+++ tests/functional/test_run_synthesis.py
@@ -2,13 +2,14 @@ import json
 import shutil
 from pathlib import Path
 
-import h5py
 import bluepysnap
-from click.testing import CliRunner
-
-from utils import TEST_PROJ_SYNTH, cwd, load_yaml
+import h5py
 from assertions import assert_node_population_morphologies_accessible
+from click.testing import CliRunner
+from utils import TEST_PROJ_SYNTH, cwd, load_yaml
+
 from circuit_build.cli import run
+from circuit_build.constants import INDEX_FILES
 
 
 def test_synthesis(tmp_path):
@@ -16,8 +17,10 @@ def test_synthesis(tmp_path):
     data_dir = TEST_PROJ_SYNTH
 
     with cwd(tmp_path):
+
         data_copy_dir = tmp_path / data_dir.name
         shutil.copytree(data_dir, data_copy_dir)
+
         manifest = load_yaml(data_copy_dir / "MANIFEST.yaml")
         node_population_name = manifest["common"]["node_population_name"]
         edge_population_name = manifest["common"]["edge_population_name"]
@@ -26,18 +29,22 @@ def test_synthesis(tmp_path):
         runner = CliRunner()
         result = runner.invoke(
             run,
-            args + ["functional"],
+            args + ["functional", "spatial_index_segment", "spatial_index_synapse"],
             catch_exceptions=False,
         )
         assert result.exit_code == 0
 
-        assert tmp_path.joinpath("CircuitConfig").stat().st_size > 100
         assert tmp_path.joinpath("sonata/node_sets.json").stat().st_size > 100
         assert tmp_path.joinpath("sonata/circuit_config.json").stat().st_size > 100
         assert tmp_path.joinpath("start.target").stat().st_size > 100
 
+        for index_file in INDEX_FILES:
+            index_file_path = (
+                f"sonata/networks/nodes/{node_population_name}/spatial_index/{index_file}"
+            )
+            assert tmp_path.joinpath(index_file_path).stat().st_size > 100
+
         nodes_file = (tmp_path / f"sonata/networks/nodes/{node_population_name}/nodes.h5").resolve()
-        assert f"CellLibraryFile {nodes_file}" in (tmp_path / "CircuitConfig").open().read()
         assert nodes_file.stat().st_size > 100
         with h5py.File(nodes_file, "r") as h5f:
             assert f"/nodes/{node_population_name}" in h5f